Fix yarn:clean and e2e testing, rename backend/gitlab.ts
