+++ client/src/preview/util/digitalTwin.ts
@@ -5,7 +5,7 @@ import { getAuthority } from 'util/envUtil';
 import { FileState } from 'preview/store/file.slice';
 import { LibraryConfigFile } from 'preview/store/libraryConfigFiles.slice';
 import { RUNNER_TAG } from 'model/backend/gitlab/constants';
-import { BackendInterface } from 'model/backend/gitlab/gitlab';
+import { BackendInterface } from 'model/backend/gitlab/interfaces';
 import {
   isValidInstance,
   logError,
