+++ client/src/preview/util/libraryAsset.ts
@@ -1,7 +1,7 @@
 import { getAuthority } from 'util/envUtil';
 import { AssetTypes } from 'model/backend/gitlab/constants';
 import { Asset } from 'preview/components/asset/Asset';
-import { BackendInterface } from 'model/backend/gitlab/gitlab';
+import { BackendInterface } from 'model/backend/gitlab/interfaces';
 import LibraryManager from './libraryManager';
 
 class LibraryAsset {
