+++ client/test/preview/unit/util/digitalTwinConfig.test.ts
@@ -1,4 +1,4 @@
-import GitlabInstance from 'model/backend/gitlab/gitlab';
+import GitlabInstance from 'model/backend/gitlab/interfaces';
 import DigitalTwin from 'preview/util/digitalTwin';
 
 const mockApi = {
