+++ client/test/preview/unit/util/init.test.ts
@@ -38,7 +38,7 @@ const mockGitlab = {
   getTriggerToken: jest.fn(),
 };
 
-jest.mock('model/backend/gitlab/gitlab', () => ({
+jest.mock('model/backend/gitlab/interfaces', () => ({
   __esModule: true,
   default: jest.fn().mockImplementation(() => mockGitlab),
   initialGitlabInstance: mockGitlab,
